FIX: move the inline-write config override to compaction time
This was causing inline writes to be ~disabled for normal user batches,
because we generate the compaction config there now too.

For the new "compacting" path we want to allow inline writes in general,
but when parts are compacted together we still want to flush those
out... so it makes the most sense to shift the override to right before
it applies.