catalog: Optimize loading audit logs on startup (#30484)
The apply step of the catalog takes in a list of durable catalog updates
and does two things with them. It inserts/removes the updates into/from
the in-memory catalog, and it generates builtin table updates. At the
start of the apply step, updates must be sorted so that they're added in
memory in the correct order.

Audit logs are not stored in memory, so the first part of the apply step
is a no-op. Audit logs are also append only, so for many environments
they are extremely large. Sorting all the audit logs can be extremely
expensive and is not actually required.

This commit separates the loading of audit logs during startup, so that
they bypass the apply step and instead directly generate builtin table
updates.

Works towards resolving #MaterializeInc/database-issues/8384