storage: make feedback UPSERT robust against ingestion spikes
For our purposes, an ingestion spike is a large batch of updates (from
the external source) at the same timestamp that we need to ingest. And
specifically we want to make the feedback UPSERT operator robust in
terms of memory/storage usage. Below, I'll first explain the behavior of
the classic UPSERT operator and the feedback UPSERT operator as they are
before this change, and later describe the behavior of the feedback
UPSERT operator after this here change.

We differentiate between costs in terms of memory, which is strictly
talking about RAM, and storage, which is talking about upsert state
being stored in the upsert state backend. The upsert state backend can
(currently) either be storing state in memory or on disk (rocksdb).

As an side, ingestion spikes occur a) when ingesting an initial large
snapshot from the source, or b) when after a period of downtime a large
batch of updates have accumulated in the external source, all of which
will be ingested at the same timestamp.

Memory/Storage usage before this change
---------------------------------------

The root of the differences is whether or not the operator supports
emitting partial updates. If we didn't care about memory costs, we would
always wait for a timestamp to be complete (not beyond the frontier
anymore) and then process its updates. This would make it so we emit at
most one pair of updates per key (a retraction of the old value, if any,
and the addition of the new value). However, this means we have to
buffer all updates of a given timestamp, with potentially unbounded
memory requirements. The solution the classic UPSERT operator uses is to
emit _partial updates_: when ingesting updates right at the input
frontier, we go into a mode where we ingest updates at that frontier as
they come in, and keep emitting pairs of retractions/updates. The
trade-off is that additional memory costs stay low (basically at
`O(0)`), but we emit more updates to persist, which will have to be
consolidated by someone at some point. Another slight downside is that
management of upsert state becomes more complicated because we have to
keep around tombstones for deleted values: the updates can be processed
in an order that is different from the order implied by their _FromTime_
(kafka partition/offset, for kafka)

With all of the above, we can see that the classic UPSERT operator has
NO additional memory costs from an ingestion spike, and storage costs
are `O(num_keys)`.

The feedback UPSERT operator cannot "just" use this optimization: it
cannot update its state in reaction to inputs, so it cannot keep track
of state for correctly emitting those partial updates. We will see below
that the solution is the same, in the end, though we have to finesse
things a bit for it to work.

Without partial emission, the feedback UPSERT operator has to buffer a
batch of updates at the same timestamp until the timestamp is complete.
Its memory costs are therefore `O(size_batch)`. Its storage costs are
the same at `O(num_keys)`. It is _not_ robust against ingestion spikes.

Making partial emission work with feedback UPSERT
-------------------------------------------------

We have to address two things: a) partial emission of updates on the
source-ingest side, and b) re-ingesting those updates on the persist
feedback input.

For the former, we do the same as the classic UPSERT operator, with the
change that we introduce _provisional values_ in upsert state. These
values are stored for a given timestamp and it's only valid to read and
use them when processing updates at that timestamp. Along with a
provisional value we have to store the _finalized value_ for that key.
It's the value that we have ingested from the persist input and is known
globally (across concurrent ingestions) to be the current value for that
key.

The bigger change is how we ingest updates on the persist side. With
partial updates, there can be multiple pairs of retractions/additions
for a key, within a timestamp, and we have to be careful to consolidate
these. Before this change, the feedback UPSERT operator had been using
two different mechanisms for ingesting updates into upsert state: a)
`consolidate_snapshot_chunk`, and b) `multi_put`. The former is used
when ingesting a "snapshot" from persist, when starting up. It's also
what's used by the classic UPSERT operator when ingesting its previous
state on startup. After the initial snapshot, the feedback UPSERT
operator would use `multi_put`, which is what the classic operator
always uses to put values into state. In order to keep using it with
partial updates, though, we would have to (when reading from the persist
input) wait until the timestamp is complete, then consolidate the batch,
and then insert. We can't do that because it would again need
`O(size_batch)` memory.

Here, the solution is to just always use `consolidate_snapshot_chunk`,
both for ingesting the initial snapshot _and_ for ingesting updates as
we go. This is trading off a slight increase in CPU usage (because we
have to bincode encode/decode values when using the "consolidation" mode
of state), against robustness. In practise, we rarely have problems with
CPU usage for sources, but we do care a lot about memory/storage, so it
seems a good tradeoff to make. One additional benefit is that we might
have slightly lower memory/storage costs, because bincode will at least
to var-length encoding of integers.

Please see comments in the code about what is required for using this
"snapshotting" mode of upsert state at steady state, after ingesting the
snapshot. As you can see in the code, we're also renaming this from
"Snapshotting" to "Consolidating", to reflect this new usage pattern.

With these changes, the feedback UPSERT operator has `O(num_keys +
num_new_keys)` storage costs, with NO additional memory costs. When
ingesting the initial source snapshot, this reduces to
`O(num_new_keys)`, which is the same as classic UPSERT.

A note on stats
---------------

When reading the code, you will find that the way we track stats about
upsert state is complicated. I see why it is that way and therefore
wanted to give some helpful notes here, also because I lightly touch
some of these parts. It might be hard to refactor while still providing
the same stats, but that might be a worthwhile endeavour.

Most of the hard-to-maintain stats are concerned with numbers of records
kept in state and total size of the records in state.

The root of the problem is that we are doing blind writes to rocksdb and
that we are using the rocksdb merge function for consolidating multiple
updates to the same key. Both make it so that we cannot, when inserting
a value for a key into state, know whether there was a previous value
and what the size of that previous value was (if any).

If we only had the in-memory state backend, which uses a `HashMap`,
things would be easy: we'd do a `let prev = insert(key, new)` and then
update both our count and size stats based on that.

Instead, what we do is that `multi_put` requires the caller to provide
metadata about previous values existing under a key. And the caller must
guarantee that this metadata is correct and up to date. Most usage of
the upsert state APIs then looks like this:

```rust
let prev_value = state.get(key);
let prev_value_metadata = derive_metadata(prev_value);
// fiddle with prev_value and prepare a new_value
let new_value = [...]

state.put(key, Put { new_value, prev_value_metadata })
```

Plus, these APIs are always `multi_get` and `multi_put`, but one hopes
one gets the point.