compute: clean up around render plan
This commit cleans up all the code around the render plan. It updates
comments in both `flat_plan.rs` and rendering, and renames the
`FlatPlan` types to better describe their current roles. For example,
the plan isn't fully flat anymore, so calling it `FlatPlan` is
misleading.

Specifically, the following types were renamed are:
 * `FlatPlan` -> `RenderPlan`
 * `FlatPlanStep` -> `Node`
 * `FlatPlanNode` -> `Expr`