Add preliminary support for other cloud providers to orchestratord
Changes orchestratord to use the CloudProvider used by the catalog, to
reduce duplication and eliminate the possibility of incompatibilities
between them.
Adds preliminary support for other cloud providers.

Does not add any special handling of them, but accepts GCP, Azure, and
Generic cloud providers.

In the future, we may want to expand our support for these to include
their equivalent of IAM roles, to support advanced features such as
non-password-based access to persist buckets or assuming roles, but that
is all out of scope for now.