storage: report load-generator sources as running only after hydration
We need this to figure out whether a load-generator source has hydrated
or not, for 0dt upgrades. It turns out that there are load-generator
sources out there in the wild that take multiple minutes to go through
their prefix, that is before they arrive at a point where they produce
new data.

We consider them to be hydrated once their offset is beyond the
resume_offset.