adapter: Allow mz_system to grant to builtin role
Previously, we did not allow anyone to grant a privilege to a builtin
role. There are some use cases, such as exposing metrics/data to
mz_support, where it would be useful to grant additional privileges to
builtin roles. This commit addresses the issue by allowing only
mz_system to grant privileges to builtin roles.