storage: track latest reported status and report on reconnect
Before, we would only pump what the dataflows reported back to the
controller. Which would make it so that when a controller re-connects it
doesn't get status messages for the latest state.

Now, we track the latest status update, per object, and also what has
been reported, per object. And on re-connect we clear out what has been
reported.