0dt: fix caught-up check for empty write frontiers
This fixes a problem that came from the combination of two changes:

- In https://github.com/MaterializeInc/materialize/pull/30518 we
introduced a hydration check for sources
- In https://github.com/MaterializeInc/materialize/pull/30536 we
  introduced re-hydration of load-generator sources in read-only mode

Load-generator sources have a tendency to be finished, that is their
write frontier goes to `[]` (the empty antichain). In those cases,
storage will not render any ingestion dataflow and we would therefore
wait indefinitely for one to be hydrated.

The fix is to treat collections whose write frontier is the empty
antichain as caught up.