Sort results on replica, merge on envd (#30558)
Sort results on replica, merge on environmentd.

Previously, we'd sort data only on evironmentd, which would cause it to
consume more CPU than necessary. This change moves some of the sorting
to clusterd, and only leaves the last merge step on environmentd.

The PR selects a minimal approach, and leaves most of the code related
to result finishing untouched. It introduces an invariant that peek
results must always be sorted according to the finishing, anything else
will lead to undefined results. However, there's nothing that enforces
the results to be sorted with the same ordering, which is potentially
bad. Inside environmentd, it uses a simple heap to combine $k$ sorted
runs into a single permutation map.

The interfaces to `RowCollection` (`new`, `sorted_view`) now take a
`&[ColumnOrder]`, and internally the implementation picks the right
comparison function. If the column order slice is empty, it'll skip
decoding the rows and directly defer to the tiebreaker.

The PR moves the `RowCollection` type into `mz-expr`, which isn't ideal.
This is required because the `ColumnOrder` type is defined here, and
we'd like to pass it to the constructor of the type. Alternatives would
be to have a function here that passes the correct comparison function
to `RowCollection`, but that seems to be strictly worse than moving the
type.
I considered moving the type to `compute-types`, which seems a better
fit, but not all uses of `RowCollection` depend on `compute-types`. If
this is upsetting, I can think about alternatives.

This complexity for sorting on the cluster is roughly
$\frac{n}{k}\cdot\log \frac{n}{k}$, where $n$ is the total number of
result records, and $k$ the number of workers. The last merge step then
has a time complexity of $n\cdot\log k$ to combine $k$ sorted runs into
one.

Follow-up items include:
* Avoid a single `Bytes` allocation for all rows, and instead keep the
individual allocations.
* Assert that all `RowCollections` are sorted equally.
* Move the binary heap into an iterator to avoid materializing the
sorted view permutation.

### Tips to the reviewer

Don't look at individual commits.

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly
considered. ([trigger-ci for additional test/nightly
runs](https://trigger-ci.dev.materialize.com/))
- [ ] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [ ] If this PR includes major [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note),
I have pinged the relevant PM to schedule a changelog post.

---------

Signed-off-by: Moritz Hoffmann <mh@materialize.com>