Tidy `Join` logic in `EquivalencePropagation` (#30569)
We recently learned to incorporate nullability information from types
into `Equivalences`, automatically. This had a curious outcome, which is
that `Join` terms would inherit this information presented as
constraints, essentially adding the constraints that various columns and
expressions and such be non-null. This was largely fine, because we
prune away redundant constraints in subsequent passes, ... unless we
don't. Some of the conclusions were too clever for our clean-up pass,
which doesn't reason as deeply as `Equivalences`.

The fix is to make sure the nullability information is included early,
which it is by virtue of the input equivalence classes containing the
information, and to then reduce the join equivalences in light of the
input equivalences, and then *do not re-introduce the nullability
information*. The join equivalences did have access to the information,
and if it was removed in light of input equivalences, that means that
the constraints are redundant. Set the column nullability of the types
to "nullable" before minimizing the join equivalences, to avoid
re-introducing the nullability information.

This removes some nullability filters that we introduced in
https://github.com/MaterializeInc/materialize/pull/30506, which based on
the changes here I strongly believe are logically redundant, even though
it is hard for me to see the explanations. We could certainly work a bit
harder to map out the explanation, and I could spot check a few, but the
ones I did check (elsewhere in `ldbc`) matched the description above of
introducing "redundant" constraints that we couldn't then remove.

### Motivation

<!--
Which of the following best describes the motivation behind this PR?

  * This PR fixes a recognized bug.

    [Ensure issue is linked somewhere.]

  * This PR adds a known-desirable feature.

    [Ensure issue is linked somewhere.]

  * This PR fixes a previously unreported bug.

    [Describe the bug in detail, as if you were filing a bug report.]

  * This PR adds a feature that has not yet been specified.

[Write a brief specification for the feature, including justification
for its inclusion in Materialize, as if you were writing the original
     feature specification.]

   * This PR refactors existing code.

[Describe what was wrong with the existing code, if it is not obvious.]
-->

### Tips for reviewer

<!--
Leave some tips for your reviewer, like:

    * The diff is much smaller if viewed with whitespace hidden.
    * [Some function/module/file] deserves extra attention.
* [Some function/module/file] is pure code movement and only needs a
skim.

Delete this section if no tips.
-->

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly
considered. ([trigger-ci for additional test/nightly
runs](https://trigger-ci.dev.materialize.com/))
- [ ] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [ ] If this PR includes major [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note),
I have pinged the relevant PM to schedule a changelog post.