Instance config with stricter offset is compatible
Introduce InstanceConfig::compatible_with to check if two configurations
are compatible. Two configurations are considered compatible iff the
logging configuration is equal and the expiration offset of the other
configuration is stricter than or equal to the current one.

Addresses a concern with replica expiration that didn't allow us to
strengthen the value without risking replica restarts in case of envd
restarts. This doesn't solve the problem of relaxing the expiration, which
would still cause replica restarts -- there doesn't seem to be a good
alternative because the replicas aren't actually compatible.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>