source/pg: kill active reader before dropping the slot
The drop of a replication slot happens concurrently with an ingestion
dataflow shutting down, therefore there is the possibility that the slot
is still in use. We really don't want to leak the slot and not
forcefully terminating the dataflow's connection risks timing out. For
this reason we always kill the active backend and drop the slot.

Fixes database-issues#8756

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>