storage/kafka: avoid dropping rdkafka activations
`GlueConsumerContext::activate` is invoked by rdkafka's "queue nonempty
callback" and is responsible for waiting up the source operator when new
upstream data is available to ingest. It does so through a
`tokio::sync::Notify`. Previously, it would call `notify_waiters`, which
is documented to only wake up existing waiters. So in cases where the
source operator was not currently waiting (e.g. because it was still
busy processing the last batch of upstream updates), we could lose
notifications. This is avoided by using `notify_one` instead, which
stores a permit that ensures that the source operator gets immediately
woken up when it starts waiting on the `Notify` next.