catalog: Simplify shard ID generation
The catalog has functionality to deterministically generate a random
shard ID for a given environment. It used when we need a shard before
the catalog is fully opened. It is strictly worse than generating a
totally random shard because we limit the amount of randomness.

The catalog also has an un-migratable collection called "settings"
which maps string keys to string values. This collection is accessible
immediately after reading in a snapshot of the catalog, before we
finish opening the catalog. Any shard ID used after reading in the
snapshot can be generated fully randomly and stored in the settings
collection.

Two shards fit into this category:

  - The builtin migration shard.
  - The expression cache shard.

This commit adds a migration to move both those shards to the settings
collection for existing environments. New environments will generate
the shards randomly and stash the IDs in the settings collection.