FIX: expand out hollow runs for mismatched runs
This code previously didn't handle unordered hollow runs correctly:
we're supposed to break up runs in the "wrong" order into
single-batch-part chunks, but the whole hollow run would be passed along
as a single part.

The no-length-bound fix was meant to avoid the need for this change,
since we'd never spill unordered runs... but it turns out it's easy to
do and not too gross, so let's just do the right thing here.