catalog: Pass timestamps for transaction commit (#30006)
This commit adds plumbing for the catalog to accept a commit timestamp
for transactions.

In general, it is up to the adapter to provide that timestamp.
Currently, the timestamp that the adapter uses depends on the context.
During startup, when the timestamp oracle isn't always readily
available, the adapter will pass some timestamp that it knows will
be greater than or equal to the current catalog upper. Sometimes this
is a write timestamp from the timestamp oracle, other times it's the
current catalog upper. During all other times, the adapter
passes a write timestamp from the timestamp oracle.

After startup, the adapter applies the catalog upper with the timestamp
oracle to linearize all writes to the catalog that happened during
startup.

When getting a timestamp to allocate IDs in the catalog, the
adapter does not apply the received timestamps. If the adapter did
apply these timestamps, then strict serializable reads on tables might
block if the table upper is not past the applied timestamp.

Catalog transactions are able to re-use the oracle write timestamp
that is already available and used for audit log writes. The adapter
always performs a write to builtin tables immediately after catalog
transactions, which will allocate and apply a new write timestamp. So
effectively catalog transaction write timestamps are applied.

The goal of this PR is to move the catalog closer to the EpochMillis
timeline and set up the necessary infrastructure to pass timestamps
from the adapter to the catalog.

This work doesn't fully move the catalog onto the EpochMillis
timeline, the following work still remains:

  - Always apply the commit timestamp with the timestamp oracle after
    committing.
  - Add the catalog shard to the txn shard with all other tables.
    This would prevent blocking table reads when applying the commit
    timestamp. Additionally, it would ensure that the catalog is
    readable at the timestamp oracle read timestamp after a table
    write.

An alternative framing of this commit is that it actually does move the
catalog shard onto the EpochMillis timeline, but the catalog shard
exhibits eventual consistency.

Works towards resolving #MaterializeInc/database-issues/8384